requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
lxml
google-api-python-client
google-auth
